made it compile\!11

+++ JREngage/src/com/janrain/android/engage/session/JRProvider.java
@@ -35,14 +35,19 @@ import android.graphics.BitmapFactory;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.StateListDrawable;
+import android.media.MediaRecorder;
 import android.os.AsyncTask;
+import android.os.Build;
+import android.text.AndroidCharacter;
 import android.text.TextUtils;
 import android.util.Config;
 import android.util.Log;
+import android.view.View;
 import com.janrain.android.engage.JREngage;
 import com.janrain.android.engage.R;
 import com.janrain.android.engage.prefs.Prefs;
 import com.janrain.android.engage.types.JRDictionary;
+import com.janrain.android.engage.ui.ColorButton;
 
 import java.io.*;
 import java.net.MalformedURLException;
@@ -81,6 +86,7 @@ public class JRProvider implements Serializable {
     private static HashMap<String, Drawable> provider_list_icon_drawables =
             new HashMap<String, Drawable>();
 
+
     // Suppressed because this class is not expected to be serialized
     @SuppressWarnings("serial")
 	private final static HashMap<String, Integer> provider_list_icon_resources =
@@ -96,7 +102,7 @@ public class JRProvider implements Serializable {
                         put("icon_blogger", R.drawable.jr_icon_blogger);
                         put("icon_facebook", R.drawable.jr_icon_facebook);
                         put("icon_flickr", R.drawable.jr_icon_flickr);
-                        put("icon_foursquare", R.drawable.jr_icon_foursquare);
+                        //put("icon_foursquare", R.drawable.jr_icon_foursquare);
                         put("icon_google", R.drawable.jr_icon_google);
                         put("icon_hyves", R.drawable.jr_icon_hyves);
                         put("icon_linkedin", R.drawable.jr_icon_linkedin);
@@ -129,7 +135,7 @@ public class JRProvider implements Serializable {
                         put("logo_blogger", R.drawable.jr_logo_blogger);
                         put("logo_facebook", R.drawable.jr_logo_facebook);
                         put("logo_flickr", R.drawable.jr_logo_flickr);
-                        put("logo_foursquare", R.drawable.jr_logo_foursquare);
+                        //put("logo_foursquare", R.drawable.jr_logo_foursquare);
                         put("logo_google", R.drawable.jr_logo_google);
                         put("logo_hyves", R.drawable.jr_logo_hyves);
                         put("logo_linkedin", R.drawable.jr_logo_linkedin);
@@ -316,17 +322,22 @@ public class JRProvider implements Serializable {
             return r;
         }
 
+        if (Build.VERSION.RELEASE.startsWith("1.5")) {
+            // 1.5 can't handle our programmatic XHDPI resource instantiation
+            return c.getResources().getDrawable(R.drawable.jr_icon_unknown);
+        }
+
+
         try {
             String iconFileName = "providericon~" + drawableName + ".png";
 
             Bitmap icon = BitmapFactory.decodeStream(c.openFileInput(iconFileName));
             if (icon != null) {
-                //icon.setDensity(android.util.DisplayMetrics.DENSITY_XHIGH);
-
                 //Our downloaded icons are all at xhdpi, but Android 2.1 doesn't have the
                 //DENSITY_XHIGH constant defined yet.  Fortunately it does the right thing
                 //if you pass in the DPI as an int
-                icon.setDensity(320);
+                // XXX 1.5
+                //icon.setDensity(320);
             }
             else {
                 c.deleteFile(iconFileName);
@@ -334,17 +345,21 @@ public class JRProvider implements Serializable {
                 return c.getResources().getDrawable(R.drawable.jr_icon_unknown);
             }
 
-            return new BitmapDrawable(c.getResources(), icon);
+            /// XXX 1.5
+            //return new BitmapDrawable(c.getResources(), icon);
         }
         catch (FileNotFoundException e) {
             downloadIcons(c);
 
             return c.getResources().getDrawable(R.drawable.jr_icon_unknown);
         }
+
+        // XXX 1.5
+        return null;
     }
 
 
-    public Drawable getProviderListIconDrawable(Context c) {
+    public Drawable getProviderIcon(Context c) {
         return getDrawable(c,
                 "icon_" + mName,
                 provider_list_icon_drawables,
@@ -392,7 +407,7 @@ public class JRProvider implements Serializable {
             public Void doInBackground(Void... s) {
                 for (String iconFileName : iconFileNames) {
                     try {
-                        if (Arrays.asList(c.fileList()).contains("providericons~" + iconFileName))
+                        if (Arrays.asList(c.fileList()).contains("providericon~" + iconFileName))
                             continue;
 
                         Log.d(TAG, "Downloading icon: " + iconFileName);
@@ -402,7 +417,10 @@ public class JRProvider implements Serializable {
                         FileOutputStream fos = c.openFileOutput("providericon~" + iconFileName,
                                 Context.MODE_PRIVATE);
 
-                        while (is.available() > 0) fos.write(is.read());
+                        byte buffer[] = new byte[1000];
+                        int code;
+                        while ((code = is.read(buffer, 0, buffer.length)) > 0) fos.write(buffer, 0, code);
+                        //while (is.available() > 0) fos.write(is.read());
 
                         fos.close();
                     } catch (MalformedURLException e) {
